==> 08c_pGLMMresults.Rmd <==
---
title: "pGLMM results"
output: 
  html_document:
    toc: true
    toc_float: true
    collapsed: false
    toc_depth: 4
    df_print: paged
author: "Sabrina Pankey"
date: "`r Sys.Date()`"
editor_options: 
  chunk_output_type: console
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,warning=FALSE, message=FALSE)
```

```{r }
# make bubble chart for coevolution analyses
mainDir<-"~/Box Sync/MPL/Lesser DOB sponges/coevolution_analyses/pipeline4/"
setwd(mainDir)
outDir<-"08_pglmm_results/"
dir.exists(file.path(mainDir, outDir))

library(colorspace)
library(ggplot2)
library(ggrepel)
hma.col<-rainbow_hcl(12)[2]; lma.col<-rainbow_hcl(12)[7]

#do parsing on server since txt file is so large..
# dat <- read.table('mcmc_res/all_solutions_fdr.txt',header=T,sep='\t') #slow, large file
# #drop geo.otu
# 
# pos <- aggregate(dat[,'l.95..CI'],dat[,c('bacterial_group','factor')],function(x) sum(x > 0))
# colnames(pos)[colnames(pos)=='x'] <- 'positive'
# 
# neg <- aggregate(dat[,'u.95..CI'],dat[,c('bacterial_group','factor')],function(x) sum(x < 0))
# colnames(neg)[colnames(neg)=='x'] <- 'negative'
# 
# 
# tot <- aggregate(dat[,'l.95..CI'],dat[,c('bacterial_group','factor')],function(x) length(x))
# 
# colnames(tot)[colnames(tot)=='x'] <- 'total_possible'
 
# 
# merged <- merge(pos,neg)
# merged <- merge(merged,tot)

#save(merged, file="fdr_summarized.Rds")

#load("../pipeline2_arch/mcmc_res/fdr_summarized.Rds") # loads 'merged'
load(paste0(outDir,"fdr_summarized.Rds")) # loads 'merged'

merged$bacterial_group<-gsub("_subtree[[:digit:]]","",merged$bacterial_group,perl=F)
library(dplyr)
merged2<-merged %>% dplyr::group_by(bacterial_group, factor) %>% summarise(pos=sum(positive),neg=sum(negative),tot=sum(total_possible))
merged2<-as.data.frame(merged2)
#drop boring terms from plot
#merged2<-merged2[!(merged2$factor %in% c('geo.otu','otu.phy') ),]
merged2<-merged2[!(merged2$factor=='otu.phy' ),]
merged2<-droplevels(merged2)

merged2<-merged2 %>% mutate(ICC=(pos+neg)/tot) %>% select(bacterial_group,factor,ICC)#Calc ICC column


ggplot(merged2, aes(y=factor,x=bacterial_group)) + geom_point(aes(size=ICC)) + scale_size(range=c(-0.5,8)) + theme(axis.text.x = element_text(angle = 90, hjust = 1)) 

load("03_char_results/RFimportance_class.Rds")  #loads 'imp.cl' with importance scores for how predictive each phylum is for HMA or LMA classification

load("03_char_results/RFimportance_phyla.Rds")  #loads 'imp.phy' with importance scores for how predictive each phylum is for HMA or LMA classification

impscores<-rbind(imp.cl,imp.phy)
m3<-inner_join(merged2,impscores)

m3<-m3[order(m3$HMLvLMA),]

m3$bacterial_group<-factor(m3$bacterial_group,levels=unique(m3$bacterial_group))
ggplot(m3, aes(y=factor,x=bacterial_group)) + geom_point(aes(size=ICC,color=enrichedIn)) + scale_size(range=c(-1,5)) + theme(axis.text.x = element_text(angle = 90, hjust = 1)) 
ggplot(m3, aes(x=factor,y=bacterial_group)) + geom_point(aes(size=ICC,color=enrichedIn)) + scale_size(range=c(-1,5)) + theme(axis.text.x = element_text(angle = 90, hjust = 1)) 

p1<-ggplot(m3, aes(x=ICC,y=bacterial_group)) + geom_col(aes(fill=factor))+ scale_fill_brewer(palette = "Dark2")  + theme(axis.text.x = element_text(angle = 90, hjust = 1)) +theme(legend.position = c(0.9, 0.5))
p1 
m3a<-m3 %>% group_by(bacterial_group)%>% summarise(HMAra=mean(HMAra),LMAra=mean(LMAra)) %>% tidyr::gather(.,"spongetype","relabund",HMAra,LMAra)

p2<-ggplot(m3a, aes(x=relabund,y=bacterial_group)) + geom_col(aes(fill=spongetype))  + theme(axis.text.x = element_text(angle = 90, hjust = 1)) + scale_color_manual(values=c(hma.col,lma.col))
p2<-p2 +theme(axis.title.y =element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),legend.position = c(0.9, 0.5)) 
cowplot::plot_grid(p1,p2,rel_widths = c(3, 1))

m3 %>% group_by(factor,enrichedIn) %>%summarise(meanICC=mean(ICC))
m3b<-m3 %>% filter(abs(HMLvLMA) >1) 
m3b %>% group_by(factor,enrichedIn) %>%summarise(meanICC=mean(ICC))
#ggplot(m3b, aes(x=ICC,y=bacterial_group)) + geom_col(aes(fill=factor))  + theme(axis.text.x = element_text(angle = 90, hjust = 1)) 


```

## Comparing co-diversification with other microbial diversity patterns
Plotting ICCs (pGLMM model coefficient for cophylogeny) against ASV host-richness and proportion of specialists (ASVs occupying a Single-host species)
```{r scatterplots}

#plot ICC vs %endemics for each phylum/class, color by enrichedIn

load("07_divresults/ASVs_exclusive_bygroup.rds") #datp,datc contain numbers of ASVs exclusive to either sponge type and number co-occuring across types
typeSpecdf<-rbind.data.frame(datp[,-1],datc[,-1])
typeSpecdf$bacterial_group<-c(as.character(datp[,"Phylum"]),as.character(datc[,"Class"]))
typeSpecdf<-typeSpecdf %>% mutate(typeSpecdf,totASV=HMA+LMA+both) %>% select(bacterial_group,HMA,LMA,both,totASV,proportion)

out.df<-read.csv(file="07_divresults/hostdiversity_indices.csv") #indices for N,H,D,p for class/phyla, and 'proportion' for %ASVs in single sponge
names(out.df)[names(out.df)=="X"]<-"bacterial_group"
df2<-full_join(out.df,typeSpecdf)
df2<-df2 %>% mutate(onlyHMA=(HMA/totASV),onlyLMA=(LMA/totASV))
impscores<-rbind(imp.cl,imp.phy)
impscores<- impscores %>% select(enrichedIn,pval,HMAra,LMAra,HMLvLMA,bacterial_group,MeanDecreaseAccuracy)
df2<-full_join(df2,impscores)
df2<-df2 %>% select(bacterial_group,N,H,proportion,onlyHMA,onlyLMA,totASV,enrichedIn,HMLvLMA,pval,HMAra,LMAra,HMA,LMA,both)

cophylo_m2<-merged2[merged2$factor=="Host.otu.cophy",]
dfcophy<-left_join(cophylo_m2,df2)

load("01_results/microColors.rds") #commonphyla_colors,commonclass_colors
dfcophy_com<-dfcophy[dfcophy$bacterial_group %in% c(names(commonclass_colors),names(commonphyla_colors)),]
ggplot(dfcophy, aes(y=ICC,x=proportion)) + geom_point(aes(size=ICC)) + scale_size(range=c(-1,5)) + theme(axis.text.x = element_text(angle = 90, hjust = 1)) 
ggplot(dfcophy, aes(y=proportion,x=ICC)) +
  geom_point(aes(color=enrichedIn,size=totASV)) +
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)


#summary(lm(ICC~enrichedIn+totASV+proportion,data=dfcophy))

ggplot(dfcophy, aes(y=ICC,x=totASV)) +
  geom_point(aes(color=enrichedIn,size=proportion)) +
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

#summary(lm(ICC~enrichedIn*totASV,data=dfcophy))
#summary(lm(ICC~enrichedIn*proportion,data=dfcophy))

#number of sponge species hosts, phylum average
ggplot(dfcophy, aes(y=N,x=ICC)) +
  geom_point(aes(color=enrichedIn,size=totASV)) +
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

#summary(lm(ICC~enrichedIn+totASV+H,data=dfcophy))

ggplot(dfcophy, aes(y=N,x=ICC)) +
  geom_point(aes(size=onlyLMA),alpha=0.7,color=lma.col) +
  geom_point(aes(size=onlyHMA),alpha=0.5,color=hma.col) +
 
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

ggplot(dfcophy, aes(y=ICC,x=totASV)) +
  geom_point(aes(size=onlyLMA),alpha=0.7,color=lma.col) +
  geom_point(aes(size=onlyHMA),alpha=0.5,color=hma.col) +
 
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)


#SHannon diversity (evenness)
ggplot(dfcophy, aes(y=H,x=ICC)) +
  geom_point(aes(color=enrichedIn)) +
  ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

 mean(dfcophy$ICC[dfcophy$enrichedIn=="HMA"],na.rm=T)
#[1] 0.0325041
 mean(dfcophy$ICC[dfcophy$enrichedIn=="LMA"],na.rm=T)
#[1] 0.007089751
 wilcox.test(dfcophy$ICC[dfcophy$enrichedIn=="HMA"],dfcophy$ICC[dfcophy$enrichedIn=="LMA"])

 dfgeo<-merged2[merged2$factor=="geo.otu",]
 dfgeo<-left_join(dfgeo,df2) %>% select(bacterial_group,ICC,enrichedIn)
  names(dfgeo)[ names(dfgeo)=="ICC"]<-"geoICC"

 mean(dfgeo$geoICC[dfgeo$enrichedIn=="HMA"],na.rm=T)
 mean(dfgeo$geoICC[dfgeo$enrichedIn=="LMA"],na.rm=T)
 wilcox.test(dfgeo$geoICC[dfgeo$enrichedIn=="HMA"],dfgeo$geoICC[dfgeo$enrichedIn=="LMA"])
 
 dfgeo2<- left_join(dfgeo,m3[m3$factor=="geo.otu",]) %>% filter(!is.na(enrichedIn))
gp<- ggplot(dfgeo2, aes(x=HMLvLMA,y=geoICC)) +
  geom_point(aes(color=enrichedIn),size=2)  +  scale_color_manual(values=c(hma.col,lma.col)) +ylab("Geography ICC")
 gp
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

 
 ###
  dfhostotu<-merged2[merged2$factor=="Host.otu.otuphy",]
 dfhostotu<-left_join(dfhostotu,df2) %>% select(bacterial_group,ICC,enrichedIn)
  names(dfhostotu)[ names(dfhostotu)=="ICC"]<-"hoICC"

 mean(dfhostotu$hoICC[dfhostotu$enrichedIn=="HMA"],na.rm=T)
 mean(dfhostotu$hoICC[dfhostotu$enrichedIn=="LMA"],na.rm=T)
 wilcox.test(dfhostotu$hoICC[dfhostotu$enrichedIn=="HMA"],dfhostotu$hoICC[dfhostotu$enrichedIn=="LMA"])
 
 dfhostotu2<- left_join(dfhostotu,m3[m3$factor=="Host.otu.otuphy",]) %>% filter(!is.na(enrichedIn))
gp<- ggplot(dfhostotu2, aes(x=HMLvLMA,y=hoICC)) +  scale_color_manual(values=c(hma.col,lma.col)) +
  geom_point(aes(color=enrichedIn),size=2) + ylab("Host x ASV Phylogeny ICC")
 gp
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)
 ####
 
  ###
  dfhostphy<-merged2[merged2$factor=="Host.otu.hostphy",]
 dfhostphy<-left_join(dfhostphy,df2) %>% select(bacterial_group,ICC,enrichedIn)
  names(dfhostphy)[ names(dfhostphy)=="ICC"]<-"hoICC"

 mean(dfhostphy$hoICC[dfhostphy$enrichedIn=="HMA"],na.rm=T)
 mean(dfhostphy$hoICC[dfhostphy$enrichedIn=="LMA"],na.rm=T)
 wilcox.test(dfhostphy$hoICC[dfhostphy$enrichedIn=="HMA"],dfhostphy$hoICC[dfhostphy$enrichedIn=="LMA"])
 
 dfhostphy2<- left_join(dfhostphy,m3[m3$factor=="Host.otu.hostphy",]) %>% filter(!is.na(enrichedIn))
gp<- ggplot(dfhostphy2, aes(x=HMLvLMA,y=hoICC)) + 
  scale_color_manual(values=c(hma.col,lma.col)) +
  geom_point(aes(color=enrichedIn),size=2) + ylab("ASV x Host Phylogeny ICC")
gp
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)
 ####
 
cogeo<-inner_join(dfcophy,dfgeo) 
names(cogeo)[ names(cogeo)=="ICC"]<-"cophyICC"

cogeo<-cogeo %>% mutate(ratio=cophyICC/geoICC)
 mean(cogeo$ratio[cogeo$enrichedIn=="HMA"],na.rm=T)
 mean(cogeo$ratio[cogeo$enrichedIn=="LMA"],na.rm=T)
 wilcox.test(cogeo$ratio[cogeo$enrichedIn=="HMA"],cogeo$ratio[cogeo$enrichedIn=="LMA"])
wilcox.test(cogeo$cophyICC[cogeo$enrichedIn=="HMA"],cogeo$cophyICC[cogeo$enrichedIn=="LMA"])
 
ggplot(cogeo, aes(x=geoICC,y=cophyICC)) +
  geom_point(aes(color=enrichedIn))  + geom_smooth(aes(x=geoICC,y=cophyICC, colour=enrichedIn), method=lm, se=FALSE) +
 ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

```

## Scatterpie
Plotting ICC values with phylum sizes. Points show breakdown of ASVs which are found to co-occur in both HMA or LMA sponge types, or exclusively in either. 
```{r scatterpie,results="markdown"}
library(scatterpie)
dfcophy2 <-dfcophy[!is.na(dfcophy$both),]
dfcophy2 <-dfcophy2[!is.na(dfcophy2$enrichedIn),]
dfcophy2<-droplevels(dfcophy2)

gp<-ggplot() + 
  geom_scatterpie(aes(x=log(totASV), y=ICC*50, group=bacterial_group,r=0.1), data=dfcophy2,
                           cols=c("LMA","HMA","both"),color=NA,alpha=0.9) +coord_equal() +scale_fill_manual(values=c(lma.col,hma.col,"gray")) 
gp
gp +
  ggrepel::geom_text_repel(aes(x=log(totASV), y=ICC*50,label=bacterial_group), data=dfcophy2,size=2)
#summary(lm(ICC~enrichedIn*totASV,data=dfcophy2))
summary(lm(ICC~enrichedIn+totASV,data=dfcophy2))
summary(aov(ICC~enrichedIn+totASV,data=dfcophy2))
```

```{r calcMeans,eval=F}
#( incorporated into earlier script, so use HMAra LMAra and pval variables loaded in)


#calculate average relative abundance for each bacterial_group in HMA and LMA, use to quantify 'enrichedIn'
library(phyloseq)
load("01_results/psR.rds")
dat<-data.frame(sample_data(psR))
dat$X<-rownames(dat);
load("03_char_results/species_charsRF.Rds") #char
char2<-char %>% select(Species_consensus,consensusRF_phyla)
dat2<-merge(dat,char2,by="Species_consensus",all.x=T)

rownames(dat2)<-dat$X
sample_data(psR)<-dat2

phylumGlommed<-tax_glom(psR,taxrank="Phylum")
tax_table(phylumGlommed)[,"KP"]<-paste0(tax_table(phylumGlommed)[,"Kingdom"],"_",tax_table(phylumGlommed)[,"Phylum"])
keep<-sample_names(phylumGlommed)[!is.na(sample_data(phylumGlommed)$consensusRF_phyla)]
p2<-prune_samples(keep,phylumGlommed)

ps2df<-function(ps){
OTU1 = as(otu_table(ps), "matrix")
# transpose if necessary
if(taxa_are_rows(ps)){OTU1 <- t(OTU1)}
# Coerce to data.frame
OTUdf = as.data.frame(OTU1)
return(OTUdf)
}
psdf<-ps2df(p2)
phyla_pvals<-apply(psdf,2,function(p) 
  {ttest<-t.test(p~factor(sample_data(p2)$consensusRF_phyla)) 
  return(ttest$p.value)})
names(phyla_pvals)<-tax_table(p2)[,"KP"]

#merge_samples sums otu_counts, instead of averaging across sample factor.  Manually calc average for HMA and LMA
PGroup <- transform_sample_counts(p2, function(x) x / sum(x))
SampleType <- factor(sample_data(PGroup)$consensusRF_phyla)
table(SampleType)

mean_PGroup <- sapply(levels(SampleType),function(i){
  rowMeans(otu_table(PGroup)[,SampleType==i])
})

phy <- tax_table(PGroup)[rownames(mean_PGroup ),"KP"]
rownames(mean_PGroup) <- phy







cGlommed<-tax_glom(psR,taxrank="Class")
tax_table(cGlommed)[,"KPC"]<-paste0(tax_table(cGlommed)[,"Kingdom"],"_",tax_table(cGlommed)[,"Phylum"],"_",tax_table(cGlommed)[,"Class"])
keep<-sample_names(cGlommed)[!is.na(sample_data(cGlommed)$consensusRF_phyla)]
p3<-prune_samples(keep,cGlommed)
psdf<-ps2df(p3)
class_pvals<-apply(psdf,2,function(p) 
  {ttest<-t.test(p~factor(sample_data(p3)$consensusRF_phyla)) 
  return(ttest$p.value)})
names(class_pvals)<-tax_table(p3)[,"KPC"]

CGroup<-transform_sample_counts(p3,fun= function(otu) otu/sum(otu))
SampleType <- factor(sample_data(CGroup)$consensusRF_phyla)
table(SampleType)

mean_CGroup <- sapply(levels(SampleType),function(i){
  rowMeans(otu_table(CGroup)[,SampleType==i])
})

phy <- tax_table(p3)[rownames(mean_CGroup ),"KPC"]
rownames(mean_CGroup) <- phy


#combine results, add to dataframe for plotting
pval<-c(phyla_pvals,class_pvals) %>% data.frame; colnames(pval)<-"pval"
pval$bacterial_group<-rownames(pval)


meanRA<-rbind.data.frame(mean_CGroup,mean_PGroup)
colnames(meanRA)<-c("HMAra","LMAra")
meanRA$bacterial_group<-rownames(meanRA)

dfcophy3<-left_join(dfcophy,meanRA) %>% left_join(.,pval)
dfcophy3<-dfcophy3 %>% mutate(HMLvLMA=ifelse(HMAra>LMAra,HMAra*100,(-100*LMAra)))

dfcophy3<-dfcophy3 %>% mutate(enrichedIn2=ifelse(HMAra>LMAra,"HMA","LMA"))
dfcophy3$enrichedIn2<-factor(dfcophy3$enrichedIn2)
summary(aov(ICC~enrichedIn2,data=dfcophy3))

```

```{r}
summary(aov(ICC~enrichedIn,data=dfcophy2))
gp<-ggplot(dfcophy2, aes(y=ICC,x=HMLvLMA)) +
  geom_point(aes(color=enrichedIn),size=3) +
  scale_color_manual(values=c(hma.col,lma.col))
gp
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

dfcophy2<-dfcophy2 %>% mutate(pval2=ifelse(HMAra>LMAra,pval,(-1*pval)))

gp<-ggplot(dfcophy2, aes(y=ICC,x=pval2)) +
  geom_point(aes(color=enrichedIn),size=3) +
  scale_color_manual(values=c(hma.col,lma.col))
gp
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)

 gp<-ggplot(dfcophy2, aes(y=ICC,x=pval)) +
  geom_point(aes(color=enrichedIn),size=3) +
  scale_color_manual(values=c(hma.col,lma.col))

gp
gp<-gp +facet_grid(~enrichedIn)
gp + ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)



gp<-ggplot() + 
  geom_scatterpie(aes(x=pval, y=ICC, group=bacterial_group,r=0.02), data=dfcophy2,
                           cols=c("LMA","HMA","both"),color=NA,alpha=0.9) +coord_equal() +scale_fill_manual(values=c(lma.col,hma.col,"gray")) 
gp
gp +facet_wrap(~enrichedIn)
gp+  ggrepel::geom_text_repel(aes(x=pval, y=ICC,label=bacterial_group), data=dfcophy2,size=2)


dfcophy2<-dfcophy2 %>% mutate(prop_exclus=ifelse(onlyHMA>onlyLMA,onlyHMA,(-1*onlyLMA)))

 ggplot(dfcophy2, aes(y=ICC,x=prop_exclus)) +
  geom_point(data=subset(dfcophy2,prop_exclus<0), color=lma.col) +
     geom_point(data=subset(dfcophy2,prop_exclus>0), color=hma.col) +

   ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)
 
 gp<- ggplot(dfcophy2, aes(y=ICC,x=HMLvLMA)) +
  geom_point(aes(color=enrichedIn),size=2) +
  scale_color_manual(values=c(hma.col,lma.col)) + ylab("Co-phylogeny ICC")
gp
  gp+ ggrepel::geom_text_repel(aes(label=bacterial_group),size=2)
save(dfcophy2,file=paste(outDir,"ICC.cophy.df.rds"))

```
